% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flower_color_dist_matrix}
\alias{flower_color_dist_matrix}
\title{Flower color pair-wise distance matrix}
\format{
A 10x10 matrix:
\describe{
}
}
\usage{
flower_color_dist_matrix
}
\description{
A fictional matrix of the pair-wise differences in mean flower color values (primary wavelengths measured in nm) between 10 species of flowers.
}
\examples{
flower_color_dist_matrix
}
\keyword{datasets}
