% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flower_size_dist_matrix}
\alias{flower_size_dist_matrix}
\title{Flower size pair-wise distance matrix}
\format{
A 10x10 matrix:
\describe{
}
}
\usage{
flower_size_dist_matrix
}
\description{
A fictional matrix of the pair-wise differences in flower corolla diameter between 10 species of flowers.
}
\examples{
flower_size_dist_matrix
}
\keyword{datasets}
