% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flower_tube_length_dist_matrix}
\alias{flower_tube_length_dist_matrix}
\title{Flower tube length pair-wise distance matrix}
\format{
A 10x10 matrix:
\describe{
}
}
\usage{
flower_tube_length_dist_matrix
}
\description{
A fictional matrix of the pair-wise differences in flower corolla tube length (mm) between 10 species of flowers.
}
\examples{
flower_tube_length_dist_matrix
}
\keyword{datasets}
