% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicheDIM.R
\name{nicheDIM}
\alias{nicheDIM}
\title{nicheDIM}
\usage{
nicheDIM(resource.use, dist.matrix, dataset.name, file.dir)
}
\arguments{
\item{resource.use}{A dataframe consisting of proportional resource-use values for any number of resources (cols) used by a list of users (rows) with unique IDs. The first column should represent a list of IDs for resource users (e.g., a list of bat species for which dietary composition was determined), followed by any number of columns, each representing a different resource (e.g., species of moth consumed by bats) with a proportional resource use value given for each user. The sum of an individual row should equal 1.}

\item{dist.matrix}{A matrix of pairwise trait-distance or dissimilarity values which consists entirely of resources used (e.g., a pairwise colour distance dissimilarty matrix of different flowers species visited by bees), or includes all resources used (e.g., a phylogenetic dissimilarity matrix of all Lepidoptera, including species consumed by different species of bats). Values within the matrix will be converted to proportional values relative to the maximum value within the entire matrix (scale.limit = "none"), or limited only to resources used by users within a resource.use dataframe, or across a list of resource.use dataframes (scale.limit = "used resources"). The maximum scaling value can be set in scale.limit.}

\item{dataset.name}{A name for your own reference, that identifies the dataset supplied in resource.use. This will be helpful later if you intend compare different data sets.}

\item{file.dir}{A directory to save the returned dataframe to as a csv file.}
}
\value{
A dataframe with each representing a unique user (organism or entity that utilises resources) with the following columns: "dataset.name"; "UserID"; "users.count" = number of users; "resource.count" = number of resources; "resource.used.count" = number of resources used by user; "Levins" = calculated value for Levin's niche breadth as a reference; "Dist.use" = calculated value for nicheDIM index; "Dist.use.NOTi"; "Dist.use.ALL"; "Dist.resources"  = calculated value for nicheDIM assuming all resources are equally utilised (i.e., gives an indication of the total available nice breadth); "Prop.niche.space.NOTi"; "Prop.niche.space.ALL"

nicheDIM()
}
\description{
Intuitive niche-breadth estimation in multi-trait and phylogenetic dimensions.
}
