% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollinator_visits}
\alias{pollinator_visits}
\title{Pollinator visits rates to flowers}
\format{
A data frame with 5 rows and 10 variables:
\describe{
\item{A}{Proportional visit rate to flower species "A"}
\item{B}{Proportional visit rate to flower species "B"}
\item{C}{Proportional visit rate to flower species "C"}
\item{D}{Proportional visit rate to flower species "D"}
\item{E}{Proportional visit rate to flower species "E"}
\item{F}{Proportional visit rate to flower species "F"}
\item{G}{Proportional visit rate to flower species "G"}
\item{H}{Proportional visit rate to flower species "H"}
\item{I}{Proportional visit rate to flower species "I"}
\item{J}{Proportional visit rate to flower species "J"}
}
}
\usage{
pollinator_visits
}
\description{
A fictional dataset of five different pollinator species and their proportional visit rates to 10 different flower species.
}
\examples{
pollinator_visits
}
\keyword{datasets}
