% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{size_dist}
\alias{size_dist}
\title{Flower size pair-wise distance matrix}
\format{
A 10x10 matrix:
\describe{
}
}
\usage{
size_dist
}
\description{
A fictional matrix of the pair-wise differences in flower corolla diameter between 10 species of flowers.
}
\examples{
size_dist
}
\keyword{datasets}
