% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tube_length_dist}
\alias{tube_length_dist}
\title{Flower tube length pair-wise distance matrix}
\format{
A 10x10 matrix:
\describe{
}
}
\usage{
tube_length_dist
}
\description{
A fictional matrix of the pair-wise differences in flower corolla tube length (mm) between 10 species of flowers.
}
\examples{
tube_length_dist
}
\keyword{datasets}
